Factor out manual send into its own file
